+++ packages/formx/example/lib/examples/page_view.dart
@@ -18,35 +18,76 @@ void main() {
   );
 }
 
+final form = Formx(
+  initialValues: {
+    'name': 'John Doe',
+    'email': 'john.doe@example.com',
+    'street': '123 Main',
+    'number': '456',
+    'phone': '91982224111',
+    'address': {
+      'street': '123 Main',
+      'number': '456',
+    },
+    'nested': {
+      'password': '121212',
+      'confirm': '343434',
+    },
+    'nested2': {
+      'password': '1111111',
+      'confirm': '999999',
+    },
+    'school': 'Equipe',
+  },
+);
+
+class LizStore {
+  final form = Formx();
+
+  Key key(String key) => form.key(key);
+}
+
 class PageViewExample extends StatelessWidget {
   const PageViewExample({super.key});
 
   @override
   Widget build(BuildContext context) {
     // print(context.formx.indented);
-    return Form(
-      onChanged: () {
-        final state = context.formx();
-
-        // you can access the fields directly
-        final nameState = context.field('name');
-        final isNameValid = nameState.validate();
-        print("isNameValid $isNameValid");
-
-        // and modify them
-        state.fill({'name': 'John Doe'});
-        nameState.didChange('John Doe');
-      },
-      child: Card(
-        child: Scaffold(
-          body: PageView(
-            controller: PageController(keepPage: true),
-            children: const [
-              Page1(),
-              Page2(),
-              Page3(),
-            ].keepAlive(),
-          ),
+    return Card(
+      child: Scaffold(
+        body: PageView(
+          controller: PageController(keepPage: true),
+          children: const [
+            Page1(),
+            Page2(),
+            Page3(),
+          ],
+        ),
+        bottomNavigationBar: BottomNavigationBar(
+          items: const [
+            BottomNavigationBarItem(
+              icon: Icon(Icons.filter_list),
+              label: 'Page 1',
+            ),
+            BottomNavigationBarItem(
+              icon: Icon(Icons.restore),
+              label: 'Page 2',
+            ),
+            BottomNavigationBarItem(
+              icon: Icon(Icons.school),
+              label: 'Page 3',
+            ),
+          ],
+          onTap: (index) {
+            switch (index) {
+              case 0:
+                form.values = {'school': 'ESCS', 'name': 'Arthur'};
+              case 1:
+                form.reset();
+              case 2:
+                context.pageController?.jumpToPage(2);
+            }
+          },
         ),
       ),
     );
@@ -62,29 +103,33 @@ class Page1 extends StatelessWidget {
       body: Column(
         mainAxisSize: MainAxisSize.min,
         children: [
-          TextFormField(key: const Key('name')),
-          TextFormField(key: const Key('email')),
-          Form(
-            key: const Key('nested'),
-            onChanged: context.debugForm,
-            child: Column(
-              mainAxisSize: MainAxisSize.min,
-              children: [
-                TextFormField(key: const Key('password')),
-                TextFormField(key: const Key('confirm')),
-              ],
+          TextFormField(
+            key: form.key('name'),
+            validator: Validator().required(),
+            decoration: const InputDecoration(
+              labelText: 'Name',
             ),
           ),
-          Form(
-            key: const Key('nested2'),
-            onChanged: context.debugForm,
-            child: Column(
-              mainAxisSize: MainAxisSize.min,
-              children: [
-                TextFormField(key: const Key('password')),
-                TextFormField(key: const Key('confirm')),
-              ],
+          TextFormField(
+            key: form.key('email'),
+            validator: Validator().required().email(),
+            decoration: const InputDecoration(
+              labelText: 'Email',
             ),
+          ),
+          Column(
+            mainAxisSize: MainAxisSize.min,
+            children: [
+              TextFormField(key: form.key('nested.password')),
+              TextFormField(key: form.key('nested.confirm')),
+            ],
+          ),
+          Column(
+            mainAxisSize: MainAxisSize.min,
+            children: [
+              TextFormField(key: form.key('nested2.password')),
+              TextFormField(key: form.key('nested2.confirm')),
+            ],
           ),
         ],
       ),
@@ -107,8 +152,27 @@ class Page2 extends StatelessWidget {
       body: Column(
         mainAxisSize: MainAxisSize.min,
         children: [
-          TextFormField(key: const Key('street'), onSaved: print),
-          TextFormField(key: const Key('number'), onSaved: print),
+          TextFormField(
+            key: form.key('street'),
+            validator: Validator().required(),
+            decoration: const InputDecoration(
+              labelText: 'Street',
+            ),
+          ),
+          TextFormField(
+            key: form.key('number'),
+            validator: Validator().required().numeric(),
+            decoration: const InputDecoration(
+              labelText: 'Number',
+            ),
+          ),
+          TextFormField(
+            key: form.key('phone'),
+            inputFormatters: Formatter().phone.br(),
+            decoration: const InputDecoration(
+              labelText: 'Phone',
+            ),
+          ),
         ],
       ),
       floatingActionButton: FloatingActionButton(onPressed: context.nextPage),
@@ -125,13 +189,20 @@ class Page3 extends StatelessWidget {
       body: Column(
         mainAxisSize: MainAxisSize.min,
         children: [
-          TextFormField(key: const Key('school'), onSaved: print),
+          TextFormField(
+              key: form.key('school'),
+              validator: Validator().required(),
+              decoration: const InputDecoration(
+                labelText: 'School',
+              )),
         ],
       ),
       floatingActionButton: FloatingActionButton(
         onPressed: () {
           // This is a convenient [BuildContext] extension.
-          context.formx().validate();
+          if (form.validate()) {
+            print(form.values);
+          }
         },
       ),
     );