+++ packages/formx/lib/src/fields/image_form_field.dart
@@ -3,10 +3,7 @@ import 'dart:async';
 import 'package:cross_file/cross_file.dart';
 import 'package:flutter/material.dart';
 
-import '../formx_state.dart';
-import '../models/formx_setup.dart';
-import 'file_form_field.dart';
-import 'widgets/formx_field.dart';
+import '../../formx.dart';
 
 /// A [FormField] that allows the user to pick an image.
 class ImageFormField extends FileFormField {